Allow per-sample data allocation via the platforms.
In contrast to raydata the actual use of data is not specified via the
platform but determined by the algorithm and steps that communicate by
this data.

It might be consistent to manage raydata the same way and not give it
special meaning, since the psd-templates can be specialized anyway this
should be easy to do ... I'll think about it some time.